클리핑 후 PerspectiveDivision을 하도록 변경.

+++ SoftwareRenderer/Renderer/Rasterizer/BresenhamList.h
@@ -20,7 +20,7 @@ namespace Renderer
 
 	private:
 		int _x;
-		int currentIndex;
+		size_t currentIndex;
 		std::vector<Bresenham> edges;
 	};
 }
